+++ NEWS.md
@@ -1,5 +1,7 @@
 # FSA 0.8.27 ongoing
 * Now using ROxygen2  7.0.2.
+# Removed dependency on `gplots` package as it is now orphaned. Required adding `iRichColors()` internal function.
+* `lwCompPreds()`: Removed `\dots` from arguments as it was not in usage (per request from CRAN on 3-Feb-20).
 * `repeatedRows2Keep()`: Modified. Now makes comparisons as if `NA`s are regular values.
 
 # FSA 0.8.26 22-Nov-19